requirements.txt
websockets==10.4
python-binance==1.0.16
prometheus-client==0.16.0
